Use Map Task to iterate over check-ins
Rather than looping through the check-ins and checking the
`is_last_check_in` value, we can use a Map task to create multiple
parallel executions for each check-in.

These changes also include the basis for upcoming changes with the
RecordCheckIn function. In the future this function will be used for
storing boarding positions for each check-in.