Specify ResultPath in error handler
If a ResultPath is not specified, it defaults to `$` and all the input
is overridden, so we won't get the email in the SendFailureNotification
task.

I'm not sure if this InputPath here is correct or not, since we use the
same InputPath in the CheckIn function, but I wasn't able to find
anything in the documentation which indicated what exactly gets passed
from the prior function.