Fetch Authentication headers with Selenium
These changes add a new function, sw-update-headers, which runs a
headless Chrome browser to scrape the required Authentication headers
for checking in and stores them in Parameter Store. The check-in lambda
then pulls these headers from the Parameter Store on invocation and uses
them to (hopefully) check in successfully.

Right now the sw-update-headers lambda is triggered every hour, but it
could be improved to run on check-in failures or something.

I had to fight a _lot_ with the correct Python/Selenium/Chromium
versions on Lambda. This exact combination of Python 3.6 and the
Chromedriver was the only one I could get to work. So we're stuck on 3.6
until this is resolved.